split constraints into multiple files
